agent/hello-agents/Co-creation-projects/melxy1997-ColumnWriter/requirements.txt
hello-agents>=0.1.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# MCP 支持
fastmcp>=2.0.0

# 搜索功能（可选，至少安装一个）
tavily-python>=0.3.0  # 推荐
google-search-results>=2.4.2  # SerpAPI

# HelloAgents 间接依赖
huggingface_hub>=0.19.0
datasets>=2.14.0  # 预防性添加，因为通常和 hf_hub 一起使用
pandas>=2.0.0     # 数据处理常用


